backend/requirements.txt
fastapi==0.116.1
mangum==0.19.0
pydantic==2.11.7
pyrebase4==4.7.1
python-jose[cryptography]==3.3.0
graphene==3.3.0
valkey==6.1.1